Move the generic inspection code into the generic library
